# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CERTIFICATION-SIMPLON",
        "version": "1.0"
    },
    "paths": {
...
        "/api/login": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Connecter un utilisateur existant",
                "description": "Connectez-vous avec un e-mail et un mot de passe pour obtenir un jeton d'authentification",
                "operationId": "dcf90ce65219b0ad51fc32e9f7f69843",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "utilisateur@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "motdepasse"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Connexion réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "description": "Détails de l'utilisateur",
                                            "type": "object"
                                        },
                                        "authorization": {
                                            "description": "Détails d'autorisation",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                }
            }
        },
        "/api/user": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Afficher les informations de l'utilisateur connecté",
                "operationId": "40dae923bbcb1e7f55e11bb971e17aaa",
                "responses": {
                    "200": {
                        "description": "Détails de l'utilisateur connecté",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Déconnexion d'un utilisateur",
                "operationId": "2a5c5c6e21dd0583f5d700d2f707f03f",
                "responses": {
                    "200": {
                        "description": "Déconnexion réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/bloquerUser/{user}": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Bloquer un utilisateur",
                "description": "Bloque un utilisateur dans le système.",
                "operationId": "bloquerUser",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID de l'utilisateur à bloquer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "L'utilisateur a été bloqué avec succès.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "status_message": {
                                            "type": "string",
                                            "example": "L'utilisateur a été bloqué avec succès"
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": []
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Le user est déjà bloqué.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": 400
                                        },
                                        "status_message": {
                                            "type": "string",
                                            "example": "Le user est déjà bloqué."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur lors du bloquage du user.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": 500
                                        },
                                        "status_message": {
                                            "type": "string",
                                            "example": "Erreur lors du bloquage du user."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/debloquerUser/{user}": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Débloquer un utilisateur",
                "description": "Débloque un utilisateur dans le système.",
                "operationId": "debloquerUser",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID de l'utilisateur à débloquer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "L'utilisateur a été débloqué avec succès.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "status_message": {
                                            "type": "string",
                                            "example": "L'utilisateur a été débloqué avec succès."
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": []
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Le user est déjà débloqué.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": 400
                                        },
                                        "status_message": {
                                            "type": "string",
                                            "example": "Le user est déjà débloqué."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur lors du débloquage du user.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": 500
                                        },
                                        "status_message": {
                                            "type": "string",
                                            "example": "Erreur lors du débloquage du user."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/utilisateursBloques": {
            "get": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Liste des utilisateurs bloqués",
                "description": "Récupère la liste des utilisateurs bloqués dans le système.",
                "operationId": "listeUtilisateursBloques",
                "responses": {
                    "200": {
                        "description": "Liste des utilisateurs bloqués récupérée avec succès.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "status_message": {
                                            "type": "string",
                                            "example": "Liste des utilisateurs bloqués récupérée avec succès."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "example": []
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/utilisateurs": {
            "get": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Liste des utilisateurs actifs",
                "description": "Récupère la liste des utilisateurs actifs dans le système.",
                "operationId": "listeUtilisateurs",
                "responses": {
                    "200": {
                        "description": "Liste des utilisateurs actifs récupérée avec succès.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "status_message": {
                                            "type": "string",
                                            "example": "Liste des utilisateurs actifs récupérée avec succès."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "example": []
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}